+++ examples/main/shared/viewer/Viewer.jsx
@@ -56,6 +56,11 @@ export default class Viewer extends PureComponent {
     const { isMobile, initialState, locale, seoMode, localeResource } = this.props;
     const { expandModeIsOpen, expandModeIndex, disabled } = this.state;
     const viewerProps = {
+      helpers: {
+        // This is for debugging only
+        onViewerAction: async (actionName, pluginId, value) =>
+          console.log('onViewerAction', actionName, pluginId, value),
+      },
       localeResource,
       locale,
       relValue,